requirements.txt
annotated-types==0.7.0
anthropic==0.96.0
anyio==4.13.0
APScheduler==3.11.2
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
click-default-group==1.2.4
colorama==0.4.6
distro==1.9.0
docstring_parser==0.18.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
jiter==0.14.0
oauthlib==3.3.1
pluggy==1.6.0
praw==7.8.1
prawcore==2.4.0
pydantic==2.13.3
pydantic_core==2.46.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.1
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
sqlite-fts4==1.0.3
sqlite-utils==3.39
tabulate==0.10.0
tweepy==4.16.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
update-checker==0.18.0
urllib3==2.6.3
websocket-client==1.9.0
